+++ orbax/__init__.py
@@ -15,4 +15,4 @@
 """Orbax API."""
 
 # A new PyPI release will be pushed everytime `__version__` is increased.
-__version__ = '0.1.0'
+__version__ = '0.1.1'