+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+CheckpointManager `delete` API, which allows deleting an existing step.
+
 ## [0.1.0] - 2022-1-04
 
 ### Added